cleared most errors
